requirements.txt
streamlit==1.32.0
requests==2.31.0
beautifulsoup4==4.12.3
duckduckgo-search==6.2.13
Pillow==10.3.0